restructure events submodule
