+++ ref_builder/events/__init__.py
@@ -1,0 +1,1 @@
+"""Events that are emitted to the repository event log."""
